fix(core): stop dsd_state::symbolcnt overflowing after five days of uptime
symbolcnt is a free-running count of the symbols getSymbol() has produced.
It was a signed int and nothing ever bounded it, so at 4800 symbols/s the
increment in symbol_commit_symbol() reached INT_MAX after about 5.2 days and
overflowed. That is undefined behaviour, and under the asan-ubsan-debug
preset it aborts the decoder outright:

  src/dsp/dsd_symbol.c:1856:21: runtime error: signed integer overflow:
  2147483647 + 1 cannot be represented in type 'int'

Make symbolcnt a uint32_t so the increment wraps instead. Nothing reads it
as a magnitude: the datascope compares it against a small refresh period and
zeroes it, and the SPS hunt only ever differences it against its own mark.
sps_hunt_symbolcnt_mark becomes uint32_t alongside it, which is what keeps
that difference exact across the rollover -- the mark has to wrap on the same
modulus as the value it tracks.

frame_sync_sps_hunt_note_handler_consumption() was already written for a
wrapping counter: it did the subtraction through explicit unsigned casts and
used the top half of the range to spot "an unrelated subsystem zeroed
symbolcnt". Those casts are now redundant and the heuristic is unchanged, so
uint32_t rather than a wider type is what preserves it. The datascope
comparison in print_datascope() gains a cast on the int refresh period, which
opts->scoperate fixes at 320.

Tests: DSP_SYMBOL_REPLAY seeds the counter at INT32_MAX and at UINT32_MAX and
drives the real getSymbol(), which reproduces the UBSan abort against the old
field type. That abort only fails the run under `ctest --preset
asan-ubsan-debug`, the sole preset setting UBSAN_OPTIONS=halt_on_error=1;
elsewhere the diagnostic prints and the binary exits 0, so under dev-debug what
catches a revert is instead the -Werror=sign-compare break the new assertions
provoke against a signed field. The case says so.

FRAME_SYNC_INTERNAL_HELPERS gains a seam onto the consumption helper and pins
the debit across a rollover, including the reset-to-zero and below-a-frame
paths. It is a contract test rather than a regression test -- the helper's old
explicit (unsigned int) casts already made that arithmetic modular, so it
passes against the pre-change field types too -- and its comment now says that
outright.

The backwards-jump comment in that helper also gains the third site that zeroes
symbolcnt, print_datascope(), which this commit touches: it can zero the
counter faster than handlers consume, so an interval that decoded a frame can
straddle a reset and be credited nothing. The comment states why that is
accepted -- nothing in the tree sets opts->datascope to 1, and a missed credit
only leaves the hunt on its pre-#390 undebited dwell.

Closes #395.

+++ src/core/frames/dsd_dibit.c
@@ -170,7 +170,8 @@ print_datascope(const dsd_opts* opts, dsd_state* state, const float* sbuf2, int 
     }
     const float scale = 32.0f / span;
     build_datascope_spectrum(sbuf2, count, scale, spectrum);
-    if (state->symbolcnt > (4800 / opts->scoperate)) {
+    /* dsd_state::symbolcnt is unsigned; the refresh period is a small positive int. */
+    if (state->symbolcnt > (uint32_t)(4800 / opts->scoperate)) {
         dsd_call_snapshot call;
         DSD_MEMSET(&call, 0, sizeof(call));
         (void)dsd_call_state_get(state, 0U, &call);
